Move permanent MetricTask code into tasks subpackage.
