+++ doc/lsst.verify/index.rst
@@ -102,4 +102,8 @@ Python API reference
    :no-main-docstr:
    :no-inheritance-diagram:
 
+.. automodapi:: lsst.verify.tasks
+   :no-main-docstr:
+   :no-inheritance-diagram:
+
 .. _SQUASH: https://squash.lsst.codes
