+++ python/lsst/verify/__init__.py
@@ -43,4 +43,3 @@ from .blobset import *
 from .jobmetadata import *
 from .job import *
 from .output import *
-from .metricTask import *
