+++ python/lsst/verify/tasks/metricTask.py
@@ -26,11 +26,11 @@ __all__ = ["MetricComputationError"]
 class MetricComputationError(RuntimeError):
     """This class represents unresolvable errors in computing a metric.
 
-    `gen2tasks.MetricTask` raises ``MetricComputationError`` instead of
-    other data- or processing-related exceptions to let code that calls a mix
-    of data processing and metric tasks distinguish between the two.
-    Therefore, most ``MetricComputationError`` instances should be chained to
-    another exception representing the underlying problem.
+    `lsst.verify.gen2tasks.MetricTask` raises ``MetricComputationError``
+    instead of other data- or processing-related exceptions to let code that
+    calls a mix of data processing and metric tasks distinguish between
+    the two. Therefore, most ``MetricComputationError`` instances should be
+    chained to another exception representing the underlying problem.
     """
     pass
